Refector

+++ src/main/java/com/sparta/team15/service/CardService.java
@@ -2,8 +2,18 @@ package com.sparta.team15.service;
 
 import com.sparta.team15.dto.CardRequestDto;
 import com.sparta.team15.dto.CardResponseDto;
+import com.sparta.team15.entity.BoardColumn;
 import com.sparta.team15.entity.Card;
+import com.sparta.team15.entity.User;
+import com.sparta.team15.enums.MessageEnum;
+import com.sparta.team15.exception.AuthorizedException;
+import com.sparta.team15.exception.BoardColumnErrorCode;
+import com.sparta.team15.exception.NotFoundException;
+import com.sparta.team15.exception.UserErrorCode;
+import com.sparta.team15.repository.BoardColumnRepository;
+import com.sparta.team15.repository.BoardUserRepository;
 import com.sparta.team15.repository.CardRepository;
+import com.sparta.team15.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -19,12 +29,22 @@ import java.util.List;
 public class CardService {
 
     private final CardRepository cardRepository;
+    private final BoardColumnRepository boardColumnRepository;
+    private final BoardUserRepository boardUserRepository;
 
     // 카드 생성
-    public void createCard(CardRequestDto requestDto) {
+    public void createCard(CardRequestDto requestDto, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+        BoardColumn boardColumn = boardColumnRepository.findById(requestDto.getColumnId())
+                .orElseThrow(() -> new NotFoundException(BoardColumnErrorCode.NOT_FOUND_COLUMN));
+
+        boardUserRepository.findByUserIdAndBoardId(user.getId(), boardColumn.getBoard().getId())
+                .orElseThrow(() -> new NotFoundException(BoardColumnErrorCode.NOT_TEAM_MEMBER));
+
         Card card = new Card(
+                user,
                 requestDto.getAuthor(),
-                requestDto.getColumnId(),
+                boardColumn,
                 requestDto.getContent(),
                 requestDto.getDescription(),
                 requestDto.getDate()
@@ -33,7 +53,8 @@ public class CardService {
     }
 
     // 카드 전체 목록 조회
-    public List<CardResponseDto> getAllCards(int page, int size) {
+    public List<CardResponseDto> getAllCards(int page, int size, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
 
         Pageable pageable = PageRequest.of(page, size);
 
@@ -42,16 +63,20 @@ public class CardService {
     }
 
     // 카드 상태별 조회
-    public List<CardResponseDto> getCardsByStatus(Long columnId, int page, int size) {
+    public List<CardResponseDto> getCardsByStatus(Long columnId, int page, int size, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+        BoardColumn boardColumn = boardColumnRepository.findById(columnId)
+                .orElseThrow(() -> new NotFoundException(BoardColumnErrorCode.NOT_FOUND_COLUMN));
 
         Pageable pageable = PageRequest.of(page, size);
 
-        return cardRepository.findByColumnId(columnId, pageable).stream()
+        return cardRepository.findByBoardColumn(boardColumn, pageable).stream()
                 .map(CardResponseDto::new).toList();
     }
 
     // 카드 작업자별 조회
-    public List<CardResponseDto> getCardsByAuthor(String author, int page, int size) {
+    public List<CardResponseDto> getCardsByAuthor(String author, int page, int size, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
 
         Pageable pageable = PageRequest.of(page, size);
 
@@ -60,9 +85,15 @@ public class CardService {
     }
 
     // 카드 수정
-    public void updateCard(Long cardId, CardRequestDto requestDto) {
+    public void updateCard(Long cardId, CardRequestDto requestDto, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+
         Card card = cardRepository.findById(cardId)
-                .orElseThrow(() -> new IllegalArgumentException("카드를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundException(UserErrorCode.NOT_FOUND_CARD));
+
+        if (card.getUser().getId() != user.getId()) {
+            throw new AuthorizedException(UserErrorCode.NOT_AUTHORIZATION_ABOUT_CARD);
+        }
 
         card.update(requestDto.getAuthor(), requestDto.getContent(), requestDto.getDescription(), requestDto.getDate());
         cardRepository.save(card);
@@ -70,16 +101,29 @@ public class CardService {
     }
 
     // 카드 삭제
-    public void deleteCard(Long cardId) {
+    public void deleteCard(Long cardId, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
 
         Card card = cardRepository.findById(cardId)
-                .orElseThrow(() -> new IllegalArgumentException("카드를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundException(UserErrorCode.NOT_FOUND_CARD));
+
+        if (card.getUser().getId() != user.getId()) {
+            throw new AuthorizedException(UserErrorCode.NOT_AUTHORIZATION_ABOUT_CARD);
+        }
 
         cardRepository.delete(card);
     }
 
     // 카드 순서 이동
-    public void updateCardPosition(Long cardId) {
+    public void updateCardPosition(Long cardId, CardRequestDto requestDto, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+
+        Card card = cardRepository.findById(cardId)
+                .orElseThrow(() -> new NotFoundException(UserErrorCode.NOT_FOUND_CARD));
+
+        card.updatePosition(requestDto.getPosition());
+        cardRepository.save(card);
+
 
     }
